Add files via upload
Design an ATM machine that has 3 main features
1) Deposit
2) Withdraw
3) Check balance
Current balance is 5000, If the withdraw balance is equal or less than 5000 then cash out the desired
withdraw amount. If it exceed from current balance then deliver a message that you dont have
enough balance. When user select an option "1" then update the current balance. If user choose an
option, "3" then show them their current balance.
Task 7.2
GameXone offering a new position in their software house, you are selected for an interview. The
criteria for selecting your position either as a programmer or management person is that if you are
having 4-5 years of experience in game development than you will be hired as a Developer
otherwise they will shortlist your for management.
Task 7.3
Design a career counseling system for students to select their field. You are requested to think like a
counsellor who guide students to choose a career according to their interest and passion. If their
hobbies , or interest is in arts, drawing , management , suggest them a field of Arts , if they are
logical, having expertise and interest in programming , problem solving ,technologies than suggest to
pursue their career in Engineering domain . If they showed their interest in bio , science , discipline
and development of humanity than suggest them a medical field.